+++ app/src/main/kotlin/hi3/hashkit/di/AppModule.kt
@@ -45,6 +45,7 @@ object AppModule {
                 HashkitDatabase.MIGRATION_6_7,
                 HashkitDatabase.MIGRATION_7_8,
                 HashkitDatabase.MIGRATION_8_9,
+                HashkitDatabase.MIGRATION_9_10,
             )
             .build()